Move locales utils from calendar to internal
