+++ src/date-range-picker/index.tsx
@@ -3,7 +3,7 @@
 import React, { Ref, useEffect, useRef, useState } from 'react';
 import styles from './styles.css.js';
 import { DateRangePickerProps } from './interfaces';
-import { normalizeLocale } from '../calendar/utils/locales';
+import { normalizeLocale } from '../internal/utils/locale';
 import useForwardFocus from '../internal/hooks/forward-focus';
 import { KeyCode } from '../internal/keycode';
 import clsx from 'clsx';
