+++ src/internal/utils/locale/index.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { mergeLocales } from './merge-locales';
+export { normalizeLocale } from './normalize-locale';
+export { normalizeStartOfWeek } from './normalize-start-of-week';
